+++ asset-manifest.json
@@ -1,9 +1,9 @@
 {
   "files": {
-    "static/js/0.4f40e7df.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.4f40e7df.chunk.js",
-    "static/js/0.4f40e7df.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.4f40e7df.chunk.js.map",
-    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.6ccaaf2c.chunk.js",
-    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.6ccaaf2c.chunk.js.map",
+    "static/js/1.57fd28c5.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.57fd28c5.chunk.js",
+    "static/js/1.57fd28c5.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.57fd28c5.chunk.js.map",
+    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.fb0918fc.chunk.js",
+    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.fb0918fc.chunk.js.map",
     "runtime~main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js",
     "runtime~main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js.map",
     "index.php": "/react-wp/wp-content/themes/inu-v2-b/index.php"
