+++ react-src/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/react-hooks": "^3.0.0",
     "@devloco/react-scripts-wptheme": "3.0.1-wptheme-3.0.5",
     "apollo-boost": "^0.4.3",
     "flatted": "^2.0.1",
