+++ react-src/src/hooks/useQuery.js
@@ -1,43 +1,5 @@
-import { useState } from "react";
-import { useQuery as _useQuery } from "react-apollo-hooks";
-
-/**
- * Pagination for the query
- * See https://github.com/trojanowski/react-apollo-hooks/issues/182
- *
- * @param  {[type]} fetchMore The pagination handler
- * @param  {[type]} data      The data returned by the query
- * @param  {[type]} variables The query variables
- * @return {[type]}           The pagination handler and status
- */
-const useLoadMore = (fetchMore, data, variables) => {
-  const [isLoadingMore, setIsLoadingMore] = useState(false);
-
-  function loadMore() {
-    setIsLoadingMore(true);
-
-    fetchMore({
-      variables: {
-        ...variables,
-        skip: data.list.length
-      },
-      updateQuery: (previousResult, { fetchMoreResult }) => {
-        setIsLoadingMore(false);
-
-        if (!fetchMoreResult) {
-          return previousResult;
-        }
-
-        return {
-          ...fetchMoreResult,
-          edges: [...previousResult.edges, ...fetchMoreResult.edges]
-        };
-      }
-    });
-  }
-
-  return [isLoadingMore, loadMore];
-};
+import React from "react";
+import { useQuery as _useQuery } from "@apollo/react-hooks";
 
 /**
  * The query hook
@@ -51,20 +13,20 @@ const useQuery = (schema, variables = {}, options = {}) => {
   /**
    * Run the query
    */
-  const { fetchMore, data, ...other } = _useQuery(schema, {
+  const { data, loading, error } = _useQuery(schema, {
     variables,
     ...options
   });
 
   /**
-   * Handle pagination
+   * Display error
    */
-  const [isLoadingMore, loadMore] = useLoadMore(fetchMore, data, variables);
+  if (error) return <p>ERROR: {error.message}</p>;
 
   /**
    * Return the data, the pagination handlers and more
    */
-  return { isLoadingMore, loadMore, fetchMore, data, ...other };
+  return { data, loading };
 };
 
-export { useQuery, useLoadMore };
+export { useQuery };
